requirements.txt
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.6.15
distlib==0.3.5
dj-database-url==1.0.0
dj-static==0.0.6
Django==3.2.8
django-heroku==0.3.1
djangorestframework==3.13.1
filelock==3.7.1
gunicorn==20.1.0
Pillow==9.2.0
pipenv==2022.7.24
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.9.0
PyMySQL==1.0.2
pytz==2022.1
sqlparse==0.4.2
static3==0.7.0
toml==0.10.2
tzdata==2022.1
virtualenv==20.16.2
virtualenv-clone==0.5.7
whitenoise==6.2.0
xlwt==1.3.0
